pyproject.toml
[project]
name = "teacher-lang"
version = "0.2.0"
description = "AI-powered student answer grading pipeline"
readme = "README.md"
requires-python = ">=3.14"

dependencies = [
    # LangGraph workflow
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",

    # LLM providers
    "anthropic>=0.40.0",           # Claude – grading
    "google-generativeai>=0.8.0",  # Gemini – OCR / vision

    # Document handling
    "python-docx>=1.1.0",

    # Utilities
    "python-dotenv>=1.0.0",
    "pillow>=10.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"